MYBEAUTIP-791 동영상 목록 조회 API 수정

Approved-by: KimJaeHun
Approved-by: Tekhun Jeong

+++ src/main/java/com/jocoos/mybeautip/domain/video/persistence/domain/VideoCategory.java
@@ -18,10 +18,10 @@ public class VideoCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(insertable = false, updatable = false)
-    public Long id;
+    public Integer id;
 
     @Column
-    public Long parentId;
+    public Integer parentId;
 
     @Enumerated(EnumType.STRING)
     private VideoCategoryType type;
